+++ docs/scripts/services/date.service.js
@@ -16,10 +16,11 @@
 
 	dateService.$inject = [
 		'moment',
-		'logService'
+		'logService',
+		'appConstant'
 	];
 
-	function dateService(moment, logService) {
+	function dateService(moment, logService, appConstant) {
 		const data = {
 			service : 'dateService',
 			days    : 21,
@@ -36,7 +37,8 @@
 			isDay,
 			isWeekend,
 			setNextDayAfterWeekend,
-			weekendAndExceptionsStuff
+			weekendAndExceptionsStuff,
+			readable
 		};
 
 		return {
@@ -48,10 +50,13 @@
 			if ($date) {
 
 				// Convert the date
-				let date = moment($date);
+				logService.service(data.service, 'original date is: ' + $date);
+				let date = moment($date, 'EEEE MM MMMM yyyy');
+				logService.service(data.service, 'moment original date is: ' + readable(date));
 
 				// Add 21 days
 				date = methods.add(date, data.days, 'days');
+				logService.service(data.service, 'new +21 days date is: ' + readable(date));
 
 				return methods.weekendAndExceptionsStuff(date);
 			}
@@ -95,6 +100,7 @@
 
 				// Set the next monday
 				date = methods.setNextDayAfterWeekend(date);
+				logService.service(data.service, 'new date after weekend: ' + readable(date));
 			}
 
 			// Check if this is an fr exception
@@ -103,9 +109,14 @@
 
 				// Add one day
 				date = methods.add(date, data.one, 'days');
+				logService.service(data.service, 'new date after exception day: ' + readable(date));
 				date = weekendAndExceptionsStuff(date);
 			}
 			return date;
+		}
+
+		function readable($date) {
+			return moment($date).format(appConstant.moment.readableFormat);
 		}
 	}